Removed manual updates for 35-rules-desc.conf
The complexity of updating this config file has lead to several bugs being opened.  We can leave the file in the stream as it will only add one default rule and the stock file works. 

Also, the tracking ID is not part of OPNsense so this only works for pfSense.